+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 0.0-18 (2019-06-11 / a27f7fc)
 
 ## Added